Remove joint support
